web/requirements.txt
rich
colorama
python-dotenv
